Architectural change + groups
Got rid of the MessageFactory class and its interface. Instead, a range of mapper classes have been defined in the data access layer. This is a preliminary step to separating model logic and model states, which should make the code cleaner.

Committed change that was somehow left out of the previous commit.